+++ setup.h
@@ -3,13 +3,13 @@
 
 #define ARRAY_SIZE(arr) (sizeof(arr) / sizeof((arr)[0]))
 
-extern const int buttons[];
-extern const int relays[];
-extern const int pwms[];
+extern const uint8_t buttons[];
+extern const uint8_t relays[];
+extern const uint8_t pwms[];
 
-extern const int NUM_BUTTONS;
-extern const int NUM_RELAYS;
-extern const int NUM_PWMS;
+extern const uint8_t NUM_BUTTONS;
+extern const uint8_t NUM_RELAYS;
+extern const uint8_t NUM_PWMS;
 
 #define PRESSED 0
 #define RELEASED 1
